Don't error out when finding a ARL we cannot parse.
CardOS 5 supports elaborate ARLs with OR and AND constructions which the
driver cannot currently parse (and even if it could, it's unclear to me
whether they would be understood by OpenSC). Upon encountering such
ARLs, do not error out, as the SELECT command itself succeeded and the
underlying object may be successfully used without a complete
understanding of ARL specifics.

Problem reported by Christian Findler.