Don't error out if we are given a pin with BACKTRACK_PIN set.
Some implementations create PINs with this bit set. This is the case with
Atos Trust Center on Windows using Card OS API 5.2B. We don't do that in
OpenSC, but this is no reason to error out when stumbling across such
PINs on cards formatted elsewhere.

Problem reported by Christian Findler.