Elliptic curve management for CardOS 5.
Keep a file on the card with a list of configured curves. Before a key
is generated, inspect this file and see if the corresponding curve has
been configured. In case it hasn't, extract the curve parameters from
OpenSSL and configure the curve on the fly. This allows keys to be
generated using only pkcs15-init (a la RSA), as ${DEITY} intended.

Add a -L option to cardos5-tool to list configured curves:

$ /var/tmp/cardos5-tool -L
Using reader with a card: CCID Compatible
prime256v1/06:08:2a:86:48:ce:3d:03:01:07
brainpoolP256r1/06:09:2b:24:03:03:02:08:01:01:07
brainpoolP512r1/06:09:2b:24:03:03:02:08:01:01:0d